Change the rest of the KilroyC desktop fonts
Next: KilroyC web fonts.

As an aside, the “1” sometimes looks too small, but if we make it any
taller it starts to look too tall. Whether is looks OK or too small
depends on the size of the font, and the only way to fix it for all
sizes is to use hinting, which is a bridge too far right now.