Update KilroyCWeb to have slightly higher “1”
Again, this is not perfect; the “1” still looks slightly lower at smaller
font sizes, but if I make it any taller, it will be too tall at other font
sizes. The only solution to make it better at this point is hinting, and
ttfqutohint has issues with squashing the caps, making them like small
caps, and until I can resolve *that*, this is the best compromise.