In coverage charts, swap “—” and “•”
This is because \11 is the very rarely used vertical tab character,
which is considered a whitespace character. By making “—” instead of
“•” use this code point, we use a glyph which browsers consider
a character which can break words instead of “•” which browsers
won’t use to break words.  This way, the “%s” regex will always
use a character which can break a word.

(The other option was to remove “©” which is the least used Xascii
character, but this feels better to me)